Rename details-under-link to hidden-details and fix test
