+++ app/assets/javascripts/discourse/app/components/hidden-details.js
