+++ app/assets/javascripts/discourse/app/templates/components/hidden-details.hbs
